extracted createShowCardsPageContext to a seperate file
