+++ app/contextCreator/createShowCardsPageContext.test.js
@@ -1,0 +1,92 @@
+import { createShowCardPageContext } from './createShowCardsPageContext';
+
+const dummySolutionData = (id, name, summary, capabilities, marketingData) => ({
+  id,
+  name,
+  summary,
+  organisation: {
+    id: '235F7D1A',
+    name: 'Halls',
+  },
+  marketingData,
+  capabilities,
+});
+
+describe('createShowCardPageContext', () => {
+  it('should create a context for 2 solutions', () => {
+    const expectedContext = {
+      solutions: [
+        {
+          id: '00001',
+          name: 'The first solution',
+          sections: [
+            {
+              id: 'summary-section',
+              name: 'Summary',
+              value: 'The summary of the first solution',
+              showTitle: true,
+            },
+          ],
+        },
+        {
+          id: '00002',
+          name: 'The second solution',
+          sections: [
+            {
+              id: 'summary-section',
+              name: 'Summary',
+              value: 'The summary of the second solution',
+              showTitle: true,
+            },
+          ],
+        },
+      ],
+    };
+
+    const twoSolutions = [dummySolutionData('00001', 'The first solution', 'The summary of the first solution'),
+      dummySolutionData('00002', 'The second solution', 'The summary of the second solution'),
+    ];
+
+    const context = createShowCardPageContext(twoSolutions);
+
+    expect(context).toEqual(expectedContext);
+  });
+
+  it('should add the capabilities context if capability data is provided', () => {
+    const expectedContext = {
+      solutions: [
+        {
+          id: '00001',
+          name: 'The first solution',
+          sections: [
+            {
+              id: 'summary-section',
+              name: 'Summary',
+              value: 'The summary of the first solution',
+              showTitle: true,
+            },
+          ],
+        },
+      ],
+      capabilities: [
+        {
+          text: 'Capability One',
+          value: 'C1',
+          checked: false,
+        },
+      ],
+    };
+
+    const capabilitiesData = [{ id: 'C1', name: 'Capability One' }];
+
+    const oneSolution = [
+      dummySolutionData('00001', 'The first solution', 'The summary of the first solution'),
+    ];
+
+    const context = createShowCardPageContext(
+      oneSolution, capabilitiesData, undefined,
+    );
+
+    expect(context).toEqual(expectedContext);
+  });
+});
